test: make gatsby-worker test adjustment global
